feat: [pss-fun-automations-007] - Close delivery loop through merge
Restore factory_visualization boundary test helpers after main merge regressed shared imports and introduced undefined projectionStub.
