+++ pkg/services/factory_visualization/import_boundary_helpers_test.go
@@ -1,0 +1,36 @@
+package factory_visualization_test
+
+import (
+	"os/exec"
+	"strings"
+	"testing"
+)
+
+const (
+	modulePrefix             = "github.com/portpowered/infinite-you/"
+	factoryVisualizationRoot = modulePrefix + "pkg/services/factory_visualization"
+	factoryRuntimeRoot       = modulePrefix + "pkg/services/factory_runtime"
+	recordingsRoot           = modulePrefix + "pkg/services/recordings"
+)
+
+func listFactoryVisualizationPackages(t *testing.T) []string {
+	t.Helper()
+
+	cmd := exec.Command("go", "list", factoryVisualizationRoot+"/...")
+	output, err := cmd.CombinedOutput()
+	if err != nil {
+		t.Fatalf("go list factory_visualization packages: %v\n%s", err, output)
+	}
+	return strings.Fields(string(output))
+}
+
+func shortFactoryVisualizationPackageName(packagePath string) string {
+	if strings.HasPrefix(packagePath, factoryVisualizationRoot) {
+		rest := strings.TrimPrefix(packagePath, factoryVisualizationRoot)
+		if rest == "" {
+			return "factory_visualization"
+		}
+		return strings.TrimPrefix(rest, "/")
+	}
+	return packagePath
+}
