+++ pkg/services/factory_visualization/recordings_import_boundary_test.go
@@ -6,12 +6,6 @@ import (
 	"testing"
 )
 
-const (
-	modulePrefix             = "github.com/portpowered/infinite-you/"
-	factoryVisualizationRoot = modulePrefix + "pkg/services/factory_visualization"
-	recordingsRoot           = modulePrefix + "pkg/services/recordings"
-)
-
 // TestProductionPackagesImportRecordingsRootOnly seals CUT-VIS-REC story 001:
 // Factory Visualization production packages may depend on Recordings only through
 // the service root contract, not nested Recordings implementation paths.
@@ -25,17 +19,6 @@ func TestProductionPackagesImportRecordingsRootOnly(t *testing.T) {
 			assertProductionImportsUseRecordingsRootOnly(t, pkg)
 		})
 	}
-}
-
-func listFactoryVisualizationPackages(t *testing.T) []string {
-	t.Helper()
-
-	cmd := exec.Command("go", "list", factoryVisualizationRoot+"/...")
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		t.Fatalf("go list factory_visualization packages: %v\n%s", err, output)
-	}
-	return strings.Fields(string(output))
 }
 
 func assertProductionImportsUseRecordingsRootOnly(t *testing.T, packagePath string) {
@@ -63,15 +46,4 @@ func isForbiddenVisualizationRecordingsImport(importPath string) bool {
 		return false
 	}
 	return strings.HasPrefix(importPath, recordingsRoot+"/")
-}
-
-func shortFactoryVisualizationPackageName(packagePath string) string {
-	if strings.HasPrefix(packagePath, factoryVisualizationRoot) {
-		rest := strings.TrimPrefix(packagePath, factoryVisualizationRoot)
-		if rest == "" {
-			return "factory_visualization"
-		}
-		return strings.TrimPrefix(rest, "/")
-	}
-	return packagePath
 }
